refactor(transport/gins): change Result.Total field type to int32
- Modify the Total field in Result struct from int64 to int32
- Update the assignment of Total in the NewResult function to use int32 directly

+++ transport/gins/result.go
@@ -23,7 +23,7 @@ import (
 type Result struct {
 	pagination.UnimplementedResponder `json:"-"`
 	Success                           bool           `json:"success"`
-	Total                             int64          `json:"total,omitempty"`
+	Total                             int32          `json:"total,omitempty"`
 	Data                              any            `json:"data,omitempty"`
 	Extra                             any            `json:"extra,omitempty"`
 	Error                             *httperr.Error `json:"error,omitempty"`
@@ -85,7 +85,7 @@ func ResultPage(c *gin.Context, resp pagination.Responder, args ...map[string]an
 		Success: true,
 		Data:    resp.GetData(),
 		Extra:   extra,
-		Total:   int64(resp.GetTotal()),
+		Total:   resp.GetTotal(),
 	})
 }
